[IIIF-843] Update manifest serialization order (#112)
This pulls in changes made in the upstream library. Those changes
can be seen at:

https://github.com/ksclarke/jiiify-presentation/commit/5c4939c2fc391ae5e4ba18df2b43564bf09e7a79

They include: updating the jiiify-presentation parent project
version, updating Javadocs that the new parent project Checkstyle
rules said needed to be updated, updating the project's pmd-rules
configuration, and specifying an order of manifest properties that
should be used on serialization. A simple test is included to
confirm that viewingHint and viewingDirection's serialization
orders have been fixed.