[IIIF-1107] Fix GPG key installation in release build
The main thing this commit does is add the following to the release build:
* gpg_private_key
* gpg_passphrase

I took the opportunity though to also clean up the release.yml a bit, giving
up on the hope of a generic release.yml that would work for all projects (at
least for now).

This also meant moving parameters from the BUILD_PROPERTIES arg to individual
properties in the release.yml's Maven configuration. The functionality is the
same, but perhaps it's clearer broken out.